Fix unauthorised count calls
