+++ src/app/core/components/layout/header/portal-navigation/portal-navigation.component.html
@@ -12,7 +12,7 @@
               >Search</a
             >
           </li>
-          @if (isTranscriber) {
+          @if (userService.isTranscriber()) {
             <li class="moj-primary-navigation__item">
               <a
                 id="your-work-link"
@@ -21,7 +21,7 @@
                 routerLinkActive="active"
                 ariaCurrentWhenActive="page"
                 >Your work
-                @if (assignedTranscriptCount) {
+                @if (assignedTranscriptCount$ | async; as assignedTranscriptCount) {
                   <span id="assignedTranscriptCount" class="moj-notification-badge"
                     ><span class="govuk-visually-hidden">Transcript requests assigned count: </span
                     >{{ assignedTranscriptCount }}</span
@@ -37,7 +37,7 @@
               routerLinkActive="active"
               ariaCurrentWhenActive="page"
               >Your audio
-              @if (unreadAudioCount) {
+              @if (unreadAudioCount$ | async; as unreadAudioCount) {
                 <span id="unreadAudioCountNotifications" class="moj-notification-badge"
                   ><span class="govuk-visually-hidden">Ready requests not viewed count: </span
                   >{{ unreadAudioCount }}</span
@@ -45,7 +45,7 @@
               }
             </a>
           </li>
-          @if (isJudge || isApprover || isRequester) {
+          @if (userService.isJudge() || userService.isApprover() || userService.isRequester()) {
             <li class="moj-primary-navigation__item">
               <a
                 class="moj-primary-navigation__link"
@@ -56,7 +56,7 @@
               >
             </li>
           }
-          @if (isTranscriber) {
+          @if (userService.isTranscriber()) {
             <li class="moj-primary-navigation__item">
               <a
                 id="transcript-requests-link"
@@ -65,7 +65,7 @@
                 routerLinkActive="active"
                 ariaCurrentWhenActive="page"
                 >Transcript requests
-                @if (unassignedTranscriptCount) {
+                @if (unassignedTranscriptCount$ | async; as unassignedTranscriptCount) {
                   <span id="unassignedTranscriptCount" class="moj-notification-badge"
                     ><span class="govuk-visually-hidden">Transcript requests unassigned count: </span
                     >{{ unassignedTranscriptCount }}</span
